requirements.txt
lyricsgenius
argparse
textblob
tqdm
sumy
Pillow
requests
numpy
